Fix #24: Add Cluster.Join() operation to join an existing cluster
Join() takes a peer's multiaddress. It then sends a "PeerAdd()"
RPC request with that peer and the local multiaddress by which
the remote peer is reached. The the remote peer uses the PeerAdd()
functionality to a) Tell the rest of the cluster about the new member
and b) send us the list of cluster members.

Join() has no REST API support, but it is provided instead via
"ipfs-cluster-service --join". A "--leave" flag has been included too,
by which a peer triggers "PeerRemove()" on itself when shutting down.

This allows a node to join an existing cluster and leave it by the
end of its life.

License: MIT
Signed-off-by: Hector Sanjuan <hector@protocol.ai>