Fix #24: Add Cluster.Join() operation to join an existing cluster
This reworks PeerAdd() and adds a Join() operation.

Join(multiaddress) uses the given bootstrap multiaddress to register
itself with it and fetch its list of cluster peers. Then it proceeds
to register itself with every peer and fetch their lists of cluster peers.
If a cluster peer is not known, then we register ourselves with it again
and so on.

PeerAdd() is now "simply" registering the new node and sending a
Join() request to it. The new node then will fetch the list
of peers and register itself with everyone like explained above.

License: MIT
Signed-off-by: Hector Sanjuan <hector@protocol.ai>