api/types: do not print some errors
The rest/client does things like "return id.ToID(), err" which is nice.
But when there's an error the ID has Cid set to nil etc. This is now
causing error messages on the console which are very misleading because
they do not correspond to the actual error of the request.

License: MIT
Signed-off-by: Hector Sanjuan <code@hector.link>