Copyedit in README
There was a slight grammar mistake. Now it is gone, and this sentence matches the one on our website.

I submitted this PR yesterday, but it wasn't signed properly. I'm opening a new PR to test where the signature should go in the GH web interface (clearly did not have it in the right place last time).

License: MIT
Signed-off-by: Michelle Hertzfeld mhertzfeld@gmail.com